Conflicts:
	drake/systems/framework/primitives/test/CMakeLists.txt

+++ drake/systems/framework/state_supervector.h
@@ -8,7 +8,7 @@
 #include <vector>
 
 #include "drake/systems/framework/state_vector.h"
-#include "drake/systems/framework/vector_interface.h"
+#include "drake/systems/framework/vector_base.h"
 
 namespace drake {
 namespace systems {
@@ -71,7 +71,8 @@ class StateSupervector : public StateVector<T> {
                                      lookup_table_.end(), index);
 
     // Use the lookup result to identify the subvector that contains the index.
-    const int subvector_id = std::distance(lookup_table_.begin(), it);
+    const int subvector_id =
+        static_cast<int>(std::distance(lookup_table_.begin(), it));
     StateVector<T>* subvector = vectors_[subvector_id];
 
     // The item at index 0 in vectors_[subvector_id] corresponds to index
